+++ banyand/observability/service.go
@@ -24,13 +24,15 @@ import (
 	"sync"
 	"time"
 
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/collectors"
 	"github.com/robfig/cron/v3"
-	"google.golang.org/grpc"
 
 	"github.com/apache/skywalking-banyandb/api/common"
 	"github.com/apache/skywalking-banyandb/banyand/metadata"
 	"github.com/apache/skywalking-banyandb/banyand/queue"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
+	"github.com/apache/skywalking-banyandb/pkg/meter"
 	"github.com/apache/skywalking-banyandb/pkg/meter/native"
 	"github.com/apache/skywalking-banyandb/pkg/run"
 	"github.com/apache/skywalking-banyandb/pkg/timestamp"
@@ -44,8 +46,8 @@ const (
 var (
 	_ run.Service = (*metricService)(nil)
 	_ run.Config  = (*metricService)(nil)
-	// MetricsServerInterceptor is the function to obtain grpc metrics interceptor.
-	MetricsServerInterceptor func() (grpc.UnaryServerInterceptor, grpc.StreamServerInterceptor) = emptyMetricsServerInterceptor
+
+	obScope = RootScope.SubScope("observability")
 )
 
 // Service type for Metric Service.
@@ -55,7 +57,7 @@ type Service interface {
 }
 
 // NewMetricService returns a metric service.
-func NewMetricService(metadata metadata.Repo, pipeline queue.Client, nodeType string, nodeSelector native.NodeSelector) Service {
+func NewMetricService(metadata metadata.Repo, pipeline queue.Client, nodeType string, nodeSelector native.NodeSelector) MetricsRegistry {
 	return &metricService{
 		closer:       run.NewCloser(1),
 		metadata:     metadata,
@@ -66,17 +68,21 @@ func NewMetricService(metadata metadata.Repo, pipeline queue.Client, nodeType st
 }
 
 type metricService struct {
-	l            *logger.Logger
-	svr          *http.Server
-	closer       *run.Closer
-	scheduler    *timestamp.Scheduler
-	metadata     metadata.Repo
-	pipeline     queue.Client
-	nodeSelector native.NodeSelector
-	listenAddr   string
-	nodeType     string
-	modes        []string
-	mutex        sync.Mutex
+	metadata         metadata.Repo
+	nodeSelector     native.NodeSelector
+	pipeline         queue.Client
+	svr              *http.Server
+	l                *logger.Logger
+	closer           *run.Closer
+	scheduler        *timestamp.Scheduler
+	nCollection      *native.MetricCollection
+	promReg          *prometheus.Registry
+	schedulerMetrics *SchedulerMetrics
+	npf              nativeProviderFactory
+	listenAddr       string
+	nodeType         string
+	modes            []string
+	mutex            sync.Mutex
 }
 
 func (p *metricService) FlagSet() *run.FlagSet {
@@ -105,13 +111,13 @@ func (p *metricService) Validate() error {
 
 func (p *metricService) PreRun(ctx context.Context) error {
 	p.l = logger.GetLogger(p.Name())
-	p.mutex.Lock()
-	defer p.mutex.Unlock()
 	if containsMode(p.modes, flagPromethusMode) {
-		MetricsServerInterceptor = promMetricsServerInterceptor
+		p.promReg = prometheus.NewRegistry()
+		p.promReg.MustRegister(collectors.NewGoCollector())
+		p.promReg.MustRegister(collectors.NewProcessCollector(collectors.ProcessCollectorOpts{}))
 	}
 	if containsMode(p.modes, flagNativeMode) {
-		NativeMetricCollection = native.NewMetricsCollection(p.pipeline, p.nodeSelector)
+		p.nCollection = native.NewMetricsCollection(p.pipeline, p.nodeSelector)
 		val := ctx.Value(common.ContextNodeKey)
 		if val == nil {
 			return errors.New("metric service native mode, node id is empty")
@@ -123,9 +129,11 @@ func (p *metricService) PreRun(ctx context.Context) error {
 			GrpcAddress: node.GrpcAddress,
 			HTTPAddress: node.HTTPAddress,
 		}
-		NativeMeterProvider = newNativeMeterProvider(ctx, p.metadata, nodeInfo)
+		p.npf = nativeProviderFactory{
+			metadata: p.metadata,
+			nodeInfo: nodeInfo,
+		}
 	}
-	initMetrics(p.modes)
 	return nil
 }
 
@@ -136,10 +144,16 @@ func (p *metricService) Name() string {
 func (p *metricService) Serve() run.StopNotify {
 	p.mutex.Lock()
 	defer p.mutex.Unlock()
+	p.initMetrics()
 	clock, _ := timestamp.GetClock(context.TODO())
 	p.scheduler = timestamp.NewScheduler(p.l, clock)
+	p.schedulerMetrics = NewSchedulerMetrics(p.With(obScope))
 	err := p.scheduler.Register("metrics-collector", cron.Descriptor, "@every 15s", func(_ time.Time, _ *logger.Logger) bool {
 		MetricsCollector.collect()
+		metrics := p.scheduler.Metrics()
+		for job, m := range metrics {
+			p.schedulerMetrics.Collect(job, m)
+		}
 		return true
 	})
 	if err != nil {
@@ -148,11 +162,11 @@ func (p *metricService) Serve() run.StopNotify {
 	metricsMux := http.NewServeMux()
 	metricsMux.HandleFunc("/_route", p.routeTableHandler)
 	if containsMode(p.modes, flagPromethusMode) {
-		registerMetricsEndpoint(metricsMux)
+		registerMetricsEndpoint(p.promReg, metricsMux)
 	}
 	if containsMode(p.modes, flagNativeMode) {
 		err = p.scheduler.Register("native-metric-collection", cron.Descriptor, "@every 5s", func(_ time.Time, _ *logger.Logger) bool {
-			NativeMetricCollection.FlushMetrics()
+			p.nCollection.FlushMetrics()
 			return true
 		})
 		if err != nil {
@@ -184,6 +198,10 @@ func (p *metricService) GracefulStop() {
 	p.closer.CloseThenWait()
 }
 
+func (p *metricService) NativeEnabled() bool {
+	return containsMode(p.modes, flagNativeMode)
+}
+
 func (p *metricService) routeTableHandler(w http.ResponseWriter, _ *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
@@ -197,4 +215,36 @@ func containsMode(modes []string, mode string) bool {
 		}
 	}
 	return false
+}
+
+// SchedulerMetrics is the metrics for scheduler.
+type SchedulerMetrics struct {
+	totalJobsStarted   meter.Gauge
+	totalJobsFinished  meter.Gauge
+	totalTasksStarted  meter.Gauge
+	totalTasksFinished meter.Gauge
+	totalTasksPanic    meter.Gauge
+	totalTaskLatency   meter.Gauge
+}
+
+// NewSchedulerMetrics creates a new scheduler metrics.
+func NewSchedulerMetrics(factory *Factory) *SchedulerMetrics {
+	return &SchedulerMetrics{
+		totalJobsStarted:   factory.NewGauge("scheduler_jobs_started", "job"),
+		totalJobsFinished:  factory.NewGauge("scheduler_jobs_finished", "job"),
+		totalTasksStarted:  factory.NewGauge("scheduler_tasks_started", "job"),
+		totalTasksFinished: factory.NewGauge("scheduler_tasks_finished", "job"),
+		totalTasksPanic:    factory.NewGauge("scheduler_tasks_panic", "job"),
+		totalTaskLatency:   factory.NewGauge("scheduler_task_latency", "job"),
+	}
+}
+
+// Collect collects the scheduler metrics.
+func (sm *SchedulerMetrics) Collect(job string, m *timestamp.SchedulerMetrics) {
+	sm.totalJobsStarted.Set(float64(m.TotalJobsStarted.Load()), job)
+	sm.totalJobsFinished.Set(float64(m.TotalJobsFinished.Load()), job)
+	sm.totalTasksStarted.Set(float64(m.TotalTasksStarted.Load()), job)
+	sm.totalTasksFinished.Set(float64(m.TotalTasksFinished.Load()), job)
+	sm.totalTasksPanic.Set(float64(m.TotalTasksPanic.Load()), job)
+	sm.totalTaskLatency.Set(float64(m.TotalTaskLatencyInNanoseconds.Load())/float64(time.Second), job)
 }