+++ banyand/stream/metadata.go
@@ -33,10 +33,14 @@ import (
 	"github.com/apache/skywalking-banyandb/banyand/internal/storage"
 	"github.com/apache/skywalking-banyandb/banyand/metadata"
 	"github.com/apache/skywalking-banyandb/banyand/metadata/schema"
+	"github.com/apache/skywalking-banyandb/banyand/observability"
 	"github.com/apache/skywalking-banyandb/banyand/queue"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
+	"github.com/apache/skywalking-banyandb/pkg/meter"
 	resourceSchema "github.com/apache/skywalking-banyandb/pkg/schema"
 )
+
+var metadataScope = streamScope.SubScope("metadata")
 
 // SchemaService allows querying schema information.
 type SchemaService interface {
@@ -57,6 +61,7 @@ func newSchemaRepo(path string, svc *service) schemaRepo {
 			svc.metadata,
 			svc.l,
 			newSupplier(path, svc),
+			resourceSchema.NewMetrics(svc.omr.With(metadataScope)),
 		),
 	}
 	sr.start()
@@ -64,7 +69,7 @@ func newSchemaRepo(path string, svc *service) schemaRepo {
 }
 
 // NewPortableRepository creates a new portable repository.
-func NewPortableRepository(metadata metadata.Repo, l *logger.Logger) SchemaService {
+func NewPortableRepository(metadata metadata.Repo, l *logger.Logger, metrics *resourceSchema.Metrics) SchemaService {
 	r := &schemaRepo{
 		l:        l,
 		metadata: metadata,
@@ -72,6 +77,7 @@ func NewPortableRepository(metadata metadata.Repo, l *logger.Logger) SchemaServi
 			metadata,
 			l,
 			newPortableSupplier(metadata, l),
+			metrics,
 		),
 	}
 	r.start()
@@ -242,6 +248,7 @@ var _ resourceSchema.ResourceSupplier = (*supplier)(nil)
 type supplier struct {
 	metadata metadata.Repo
 	pipeline queue.Queue
+	omr      observability.MetricsRegistry
 	l        *logger.Logger
 	path     string
 	option   option
@@ -254,6 +261,7 @@ func newSupplier(path string, svc *service) *supplier {
 		l:        svc.l,
 		pipeline: svc.localPipeline,
 		option:   svc.option,
+		omr:      svc.omr,
 	}
 }
 
@@ -272,20 +280,25 @@ func (s *supplier) ResourceSchema(md *commonv1.Metadata) (resourceSchema.Resourc
 }
 
 func (s *supplier) OpenDB(groupSchema *commonv1.Group) (io.Closer, error) {
+	name := groupSchema.Metadata.Name
+	p := common.Position{
+		Module:   "stream",
+		Database: name,
+	}
+
 	opts := storage.TSDBOpts[*tsTable, option]{
 		ShardNum:                       groupSchema.ResourceOpts.ShardNum,
 		Location:                       path.Join(s.path, groupSchema.Metadata.Name),
 		TSTableCreator:                 newTSTable,
+		TableMetrics:                   s.newMetrics(p),
 		SegmentInterval:                storage.MustToIntervalRule(groupSchema.ResourceOpts.SegmentInterval),
 		TTL:                            storage.MustToIntervalRule(groupSchema.ResourceOpts.Ttl),
 		Option:                         s.option,
 		SeriesIndexFlushTimeoutSeconds: s.option.flushTimeout.Nanoseconds() / int64(time.Second),
+		StorageMetricsFactory:          s.omr.With(storageScope.ConstLabels(meter.ToLabelPairs(common.DBLabelNames(), p.DBLabelValues()))),
 	}
-	name := groupSchema.Metadata.Name
 	return storage.OpenTSDB(
-		common.SetPosition(context.Background(), func(p common.Position) common.Position {
-			p.Module = "stream"
-			p.Database = name
+		common.SetPosition(context.Background(), func(_ common.Position) common.Position {
 			return p
 		}),
 		opts)